Refactor resolvers and property readers
